+++ src/core/data/tagdata.h
@@ -20,7 +20,7 @@ struct TagData {
     bool operator!=(const TagData& other) const { return !(*this == other); }
 };
 
-inline uint qHash(const TagData& tag, uint seed = 0) {
+inline size_t qHash(const TagData& tag, size_t seed = 0) {
     return qHash(tag.id, seed);
 }
